refactor: centralize workbook loading
